+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v0.2.0 (in development)
 -----------------------
 - Added a `--format` option for setting the format of the status line
+- Added a `--refresh` option for setting how often to update the status line
 
 v0.1.0 (2025-11-02)
 -------------------